prediction_service/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
yfinance==0.2.44
pandas==2.2.3
numpy==2.1.3
google-cloud-storage==2.18.2
google-cloud-logging==3.11.3
scikit-learn==1.5.2
statsmodels==0.14.4
matplotlib==3.9.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
mlflow==2.17.2
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cpu